halium: allow fallbacking to an eventual system image available inside the rootfs on "halium" file layouts
Some distributions (e.g. hybris-mobian) ship the Android system image inside the
rootfs image in userdata, thus breaking the the assumption that if rootfs.img exists,
a matching system image in userdata is present.

This commit reworks the image search inside `identify_android_image()` so that every supported
path is tested on "halium" file layouts. ANDROID_IMAGE is determined there now too, and
stores the real full path to the image.

Note: if the system image is available both in userdata and inside the halium rootfs,
the former is preferred.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>